ft/added create and list post

+++ app/api/posts/[id]/route.ts
@@ -1,5 +1,7 @@
 import prisma from "@/lib/prismadb";
 import { NextResponse } from "next/server";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../../auth/[...nextauth]/route";
 
 export async function GET(req: Request, { params }: { params: {id:string}}){
    const id = params.id
@@ -38,6 +40,10 @@ export async function PUT(req: Request,{ params }: { params: { id: string } }) {
   
 
   export async function DELETE(req: Request,{ params }: { params: { id: string } }) {
+    const session = await getServerSession(authOptions);
+    if (!session) {
+      return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
+    }
     const id = params.id;
     try {
       const post = await prisma.post.delete({ where: { id } });
